Floor plane render pipeline optimisations + shadow unit fix
- Plane::setPlane() dirty check: skip buildMesh()/initBuffers() when
  geometry parameters are unchanged; updateFloorPlane() calls on shadow
  quality change, Z-up toggle etc. are now no-ops if nothing moved
- drawFloor(): hoist 10 shared uniforms and units 32/33 sampler bind to
  once per call; configureGroundPass() handles only the 5 per-pass values
- FloorPlane::setupTextures(): switch to bindTextureUnitCached() for
  units 0 and 10; removes redundant glBindTexture calls on cache hits
- ensureShadowMapResources() extracted from loadFloor(); restores
  GL_TEXTURE0 as active unit on exit so createGPUTextureFromImage()
  cannot steal unit 2 and kill shadows
- loadFloor(): conditional create/setPlane reuses the floor plane across
  calls; passes _floorTexRepeatS/_floorTexRepeatT (was hardcoded 1,1)

Harden floor plane update caching

+++ include/TriangleMesh.h
@@ -516,8 +516,8 @@ protected:
 	unsigned int _fallbackTexture;
 	bool _hasTextureAlpha;
 	
-	unsigned int _sMax;
-	unsigned int _tMax;
+	float _sMax;
+	float _tMax;
 
 	bool _textureBindingsDirty = true;
